[accounting] Optimized acounting API queries #49
Unfortunately, without drastically change the current implementation
it is not possible to get away with a single query.

Each call to serializer.is_valid() generates a query that is
needed to perform the uniqueness check.

With this implementation, in the best cases we have 2 queries
(the first performs the uniqueness check and the second creates
the accounting object).

Closes #49