[settings] Improve default for batch mail sender #179
Changes DJANGO_FREERADIUS_BATCH_MAIL_SENDER from "testing@localhost"
to DEFAULT_FROM_EMAIL from the Django docs. DEFAULT_FROM_EMAIL is equal
to "webmaster@localhost". Instead of setting the value directly, I
chose to use this variable because it will be changed dynamically
in the future. This change is also added to the docs.

Closes #179