[docs] Specifies user to keep the django server running #223
The docs currently assume that the user is running the django-freeradius server in the background. Since the docs did not have any instances of pointing that, this commit will make the doc mention clearly the need to keep the django-freeradius server running in the background while attempting to start the freeradius server with the config changes specified in the docs. Fixes #223